batch delete

+++ src/main/java/com/icbms/iot/rest/impl/LoRaCommandServiceImpl.java
@@ -68,6 +68,9 @@ public class LoRaCommandServiceImpl implements LoRaCommandService {
     @Autowired
     private LocationInfoMapper locationInfoMapper;
 
+    @Autowired
+    private GatewayInfoMapper gatewayInfoMapper;
+
     private static final String START_ROUND_ROBIN_URI = ":9900/api-sdm/v1/pUI";
 
     private static final String STOP_ROUND_ROBIN_URI = ":9900/api-sdm/v1/stpp";
@@ -307,6 +310,14 @@ public class LoRaCommandServiceImpl implements LoRaCommandService {
     @Override
     @Transactional
     public void deleteDevicesByProjectId(String projectId, String gatewayId) {
+        GatewayInfo gatewayInfo = gatewayInfoMapper.findById(gatewayId);
+        String gatewayIp = gatewayInfo.getIpAddress();
+        CommonResponse<List<DeviceInfoDto>> devicesResp = getDevices(gatewayIp, "");
+        List<DeviceInfoDto> devices = devicesResp.getData();
+        if(CollectionUtils.isEmpty(devices))
+            return;
+
+        List<Integer> ids = devices.stream().map(DeviceInfoDto::getApplicationId).distinct().collect(Collectors.toList());
         gatewayDeviceMapMapper.deleteByGatewayId(gatewayId);
         List<DeviceBoxInfo> boxList = deviceBoxInfoMapper.findByProjectIdList(Arrays.asList(projectId));
         List<String> boxIdList = boxList.stream().filter(Objects::nonNull).map(DeviceBoxInfo::getId).distinct().collect(Collectors.toList());
@@ -317,6 +328,9 @@ public class LoRaCommandServiceImpl implements LoRaCommandService {
         deviceAlarmInfoLogMapper.deleteByProjectId(projectId);
         deviceSwitchInfoLogMapper.deleteByProjectId(projectId);
         deviceSwitchInfoDetailLogMapper.deleteByProjectId(projectId);
+        CommonResponse<Map> deleteResp = deleteDevices(gatewayIp, ids);
+        if(!"200".equals(deleteResp.getCode()))
+            throw new IotException(ErrorCodeEnum.REMOTE_CALL_FAILED);
     }
 
     private GatewayInfo convertGateWay(String gatewayIp, Map<String, Object> map) {
